# Bank_REST/src/main/java/com/example/bankcards/dto/UserPageDTO.java
package com.example.bankcards.dto;

import lombok.Builder;

import java.util.List;

@Builder
public record UserPageDTO(
        List<UserDTO> content,
        int page,
        int size,
        long totalElements,
        int totalPages
) {
}